am_ai_demo/requirements.txt
# SAM Demo - Python Dependencies
# Generated from codebase analysis - only includes libraries not part of standard Python distribution

# Snowflake connectivity and data processing
# Used for: Database connections, Snowpark DataFrames, Cortex functions, pandas integration
# The [pandas] extra ensures compatible pandas version is installed automatically
snowflake-snowpark-python[pandas]>=1.11.0
pyyaml>=6.0

# PDF processing capabilities (optional - for PDF template analysis)
# Used for: Extracting text and structure from PDF documents to create enhanced templates
# Uncomment to enable PDF processing:
# PyPDF2>=3.0.0          # PDF text extraction
# pdfplumber>=0.9.0      # Advanced PDF parsing with table detection
# pymupdf>=1.23.0        # High-performance PDF processing (alternative)

# Installation command:
# pip install -r requirements.txt
